Renamed items to osd-ace
