+++ packages/osd-ace/package.json
@@ -1,19 +1,19 @@
 {
-  "name": "@kbn/ace",
+  "name": "@osd/ace",
   "version": "1.0.0",
   "private": true,
   "main": "./target/index.js",
   "license": "Apache-2.0",
   "scripts": {
     "build": "node ./scripts/build.js",
-    "kbn:bootstrap": "yarn build --dev"
+    "osd:bootstrap": "yarn build --dev"
   },
   "dependencies": {
     "brace": "0.11.1"
   },
   "devDependencies": {
-    "@kbn/dev-utils": "1.0.0",
-    "@kbn/babel-preset": "1.0.0",
+    "@osd/dev-utils": "1.0.0",
+    "@osd/babel-preset": "1.0.0",
     "raw-loader": "^3.1.0",
     "typescript": "4.0.2"
   }
