+++ packages/osd-ace/src/ace/modes/lexer_rules/script_highlight_rules.ts
