+++ packages/osd-ace/src/ace/modes/x_json/worker/x_json.ace.worker.js
